chore: remove the legacy outbox producer API
Both writers moved to Publish in the previous commit, leaving Append and
AppendBatch with no callers. Keeping a compiling, documented way to write to
the deprecated table invites a new producer to find it and land there, which is
the one thing the cutover was for.

The one remaining insert was the legacy relay's own test fixture, so it moves
to testenv alongside the other outbox fixtures those tests already use — all of
which get deleted together when that relay goes.

The package doc went with writer.go, which was no loss: it documented an
AppendEvent that never existed and a signal step that isn't in the repo.
