+++ server/internal/outbox/repo/copyfrom.go
@@ -9,40 +9,6 @@ import (
 	"context"
 )
 
-// iteratorForBulkInsertOutboxEntries implements pgx.CopyFromSource.
-type iteratorForBulkInsertOutboxEntries struct {
-	rows                 []BulkInsertOutboxEntriesParams
-	skippedFirstNextCall bool
-}
-
-func (r *iteratorForBulkInsertOutboxEntries) Next() bool {
-	if len(r.rows) == 0 {
-		return false
-	}
-	if !r.skippedFirstNextCall {
-		r.skippedFirstNextCall = true
-		return true
-	}
-	r.rows = r.rows[1:]
-	return len(r.rows) > 0
-}
-
-func (r iteratorForBulkInsertOutboxEntries) Values() ([]interface{}, error) {
-	return []interface{}{
-		r.rows[0].OrganizationID,
-		r.rows[0].EventType,
-		r.rows[0].Payload,
-	}, nil
-}
-
-func (r iteratorForBulkInsertOutboxEntries) Err() error {
-	return nil
-}
-
-func (q *Queries) BulkInsertOutboxEntries(ctx context.Context, arg []BulkInsertOutboxEntriesParams) (int64, error) {
-	return q.db.CopyFrom(ctx, []string{"outbox"}, []string{"organization_id", "event_type", "payload"}, &iteratorForBulkInsertOutboxEntries{rows: arg})
-}
-
 // iteratorForBulkInsertPublishOutboxEntries implements pgx.CopyFromSource.
 type iteratorForBulkInsertPublishOutboxEntries struct {
 	rows                 []BulkInsertPublishOutboxEntriesParams
